+++ R/nmr_identify_regions.R
@@ -110,3 +110,57 @@ NULL
 #' @references \url{hmdb.ca}
 #' @keywords data
 NULL
+
+#' NMR peak identification (cell samples)
+#' 
+#' Identify given regions and return a data frame with plausible assignations 
+#' in cell samples.
+#' 
+#' @return a data frame with plausible assignations.
+#'
+#' @examples 
+#' \dontrun{
+#' # We identify regions from from the corresponding ppm storaged in a vector.
+#' library(AlpsNMR)
+#' ppm_to_assign <- c(4.060960203, 3.048970634,2.405935596,
+#' 3.24146865,0.990616851,1.002075066,0.955325548)
+#' identification <- nmr_identify_regions_cell (ppm_to_assign, num_proposed_compounds = 3)
+#' }
+#' @export
+#' @family peak detection functions
+#' @family peak integration functions
+#' @param ppm_to_assign A vector with the ppm regions to assign
+#' @param num_proposed_compounds set the number of proposed metabolites in `HMDB_cell`.
+nmr_identify_regions_cell <- function(vector, ...) {
+  UseMethod("nmr_identify_regions_cell")
+}
+nmr_identify_regions_cell <- function(ppm_to_assign, num_proposed_compounds = 3){
+  HMDB_cell <- NULL
+  utils::data("HMDB_cell", package = "AlpsNMR", envir = environment())
+  output_assignation_list=HMDB_cell[NULL,]
+  
+  for(ppm in ppm_to_assign){
+    print(ppm)
+    lower_ppm_right_edge <- ppm - 0.015
+    higher_ppm_left_edge <- ppm + 0.015
+    
+    ind= intersect(which(HMDB_cell$Shift_ppm<higher_ppm_left_edge),which(HMDB_cell$Shift_ppm>lower_ppm_right_edge))
+    assignation_list <- as.data.frame(HMDB_cell[ind,])
+    message("your peak at ",ppm, " probably corresponds to ",assignation_list[1,1], ", ",assignation_list[2,1],", ", assignation_list[3,1])
+    message("")
+    output_assignation_list <-rbind(output_assignation_list,assignation_list[1:num_proposed_compounds,])
+  }
+  output_assignation_list$ppm_to_assign <- rep(ppm_to_assign,each=num_proposed_compounds)
+  output_assignation_list <- output_assignation_list[order(output_assignation_list$ppm_to_assign),]
+  return(output_assignation_list)
+}
+NULL
+
+#' The Human Metabolome DataBase multiplet table: cell metabolites normally found in NMR-based metabolomics
+#'
+#' @name HMDB_cell
+#' @docType data
+#' @references \url{hmdb.ca}
+#' @keywords data
+#' 
+NULL